Created base functionality
The Cell.h and Cell.cpp files contain headers and method implementations
of the Cell class, respectively. The Cell class contains information
about the main object of the game: the cell. The cell has two fields:
type and coordinates of the current location. Also the Cell class
contains two constructors (one by default), the other with parameter.
The constructor with parameter is needed when creating a cell in the
field class, where it is created at once in a certain section of the
field.

The Coordinate class has one private field that contains column and row
values in the matrix, as well as two constructors and method of getting
coordinates.

The file enums.h contains enum types, which are necessary for various
classes, it was decided to collect the types together, to facilitate
their control.

The files Section.h and Section.cpp contain the implementation of the
class Section. Section class has a type field, which is defined
depending on the type of item field.

The Food.h and Food.cpp files contain the implementation of the Food
class. The class contains two fields: the type of food and the energy
it restores. It also has two methods that return food type and energy.

The Field.h and Field.cpp files contain implementation of Field class
which consists of fields: size of two-dimensional vector (vector is
square, element type is Section) and the two-dimensional vector itself.
So far, only the display of the generated field is implemented.