Added ability to move cells
Added a method for getting new coordinates (getNewCoordinate) in one of
the directions (up, down, left, and right) to the Coordinate class.

Added move method to the Cell class, which takes the types of
surrounding cells as a parameter and outputs the desired move
coordinates.

The Field class adds a move method that traverses a two-dimensional
vector of sections and puts together the movement of a cell.

A new enum type, Direction, is added to the file enums.h. Containing
directions for more convenient use of the getNewCoordinate method from
the Coordinate class.